refactor: make comments a subpackage
