+++ app/components/entity_mentions/comments/__init__.py
@@ -1,0 +1,11 @@
+from .integration import (
+    entity_comment_delete_handler,
+    entity_comment_edit_handler,
+    reply_with_comments,
+)
+
+__all__ = (
+    "entity_comment_delete_handler",
+    "entity_comment_edit_handler",
+    "reply_with_comments",
+)
